feat: add navigation links to dashboard layout
